odometer_auditing/requirements.txt
pandas>==1.3.4
numpy>==1.21.3
boto3>==1.20.8
requests>==2.22.0
mysql.connector>==2.2.9
